requirements.txt
streamlit
google-genai
pandas
pdfkit
markdown